requirements.txt
Django==5.2.7
django-admin-sortable2==2.2.8
django-ckeditor==6.7.3
dotenv==0.9.9
pillow==12.0.0
PyMySQL==1.1.2
python-dotenv==1.2.0
requests==2.32.5
cryptography==46.0.3
gunicorn==23.0.0